requirements.txt
Django==5.2.6
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0
